+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add function to get the name of a stem of a file
+- Add function to get the parent directory of a file
 
 ## [0.11.0] - 2020-03-26
 ### Added